Fix issue #1049 and keeps the fix for issue #910 in tact
In issue #910 the tooltip for free actions was displaying very close to the icon button, the fix for this was removing the span that surrounded the icon button.  This caused a warning on disabled actions per issue #1049.  Adding a ternary to check whether the action is enabled or disabled, so we can add/remove the span, resolves issue #1049 while keeping the fix for issue #910 in tact.